# openapi.yaml
openapi: 3.1.0
info:
  title: JisrVOC API
  version: 1.0.0
  description: >
    Backend contract for the Jisr Voice of Customer Intelligence Platform.
    All endpoints require an authenticated session (Jisr SSO / OIDC) and are
    subject to RBAC. List endpoints are cursor-paginated. This contract is the
    shared interface between the React/TypeScript frontend and the FastAPI
    backend; generate the frontend client from this file.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /themes:
    get:
      tags: [Themes]
      summary: All active themes
      parameters:
        - { name: trend, in: query, schema: { $ref: "#/components/schemas/Trend" } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ThemeSummary" }
  /themes/{id}:
    get:
      tags: [Themes]
      summary: Theme detail incl. verbatims and linked bet
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThemeDetail" }
        "404": { description: Not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SourceType: { type: string, enum: [hubspot, zendesk, canny, jira] }
...
    Language:    { type: string, enum: [ar, en, mixed] }
    Segment:     { type: string, enum: [smb, mid_market, enterprise, government] }
    Trend:       { type: string, enum: [new, rising, stable, declining] }
    BetStatus:   { type: string, enum: [draft, in_backlog, in_discovery, in_build, shipped, declined] }
...
    CountBucket:
      type: object
      properties:
        key: { type: string }
        count: { type: integer }
...
    Verbatim:
      type: object
      properties:
        text: { type: string }
        language: { $ref: "#/components/schemas/Language" }
        customer_name: { type: string }
        source: { $ref: "#/components/schemas/SourceType" }
    ThemeSummary:
      type: object
      required: [id, name_en, trend, item_count]
      properties:
        id: { type: string, format: uuid }
        name_en: { type: string }
        description_en: { type: string }
        trend: { $ref: "#/components/schemas/Trend" }
        item_count: { type: integer }
        customer_count: { type: integer }
        vote_weight: { type: integer }
        top_segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
    ThemeDetail:
      allOf:
        - $ref: "#/components/schemas/ThemeSummary"
        - type: object
          properties:
            segment_breakdown:
              type: array
              items: { $ref: "#/components/schemas/CountBucket" }
            verbatims:
              type: array
              items: { $ref: "#/components/schemas/Verbatim" }
            linked_bet: { $ref: "#/components/schemas/BetSummary" }
    BetSummary:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: "#/components/schemas/BetStatus" }
        problem_snippet: { type: string }
        affected_segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
        est_customer_count: { type: integer }
        why_now: { type: string }
        evidence_count: { type: integer }
        theme_id: { type: string, format: uuid }
...